improve valid date and witespaces

+++ ex01/RPN.hpp
@@ -18,7 +18,7 @@
 class RPN
 {
     private:
-        std::stack<int> _stack;
+        std::stack<double> _stack;
     public:
         RPN();
         ~RPN();
